Add files via upload
Original WiFi exfiltration script. It's pretty trash, and was a result of me spending about 6 hours of pain, figuring out powershell basically for the first time, and how api's worked. It was a good project, I've learned a lot from it, and I'm only uploading this for the heck of it. It's convenient to read as every command is done line by line, and isn't condensed into a one-liner, like the final finished product is. Enjoy!